Split the get and set promises for gStatus
Get colocated with component
Set colocated with the API route
