--- src/lib/database/mongoose/index.ts
@@ -1,6 +1,0 @@
-import { User, Account, Session } from './models';
-import mongoosePromise from './connection';
-
-export default mongoosePromise;
-
-export { User, Account, Session };
